refactor(calculating): combine `calculator.py` and `calculators/` into `calculating/`
